Fixed issue #20139: Export as QueXML with debug 2 and 8.3.22 : throw error

+++ application/libraries/admin/quexmlpdf.php
@@ -1871,9 +1871,9 @@ class quexmlpdf extends pdf
         }
 
         if ($number > 26) {
-            return chr((($number - 1) / 26) + 64) . chr((($number - 1) % 26) + 65);
+            return chr((int) (($number - 1) / 26) + 64) . chr((int) (($number - 1) % 26) + 65);
         } else {
-            return chr($number + 64);
+            return chr((int) ($number + 64));
         }
     }
